Move classes to the proper packages
