+++ src/main/java/com/github/novotnyr/idea/jwt/rs256/RS256Panel.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.github.novotnyr.idea.jwt.ui.secretpanel.RS256Panel">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.github.novotnyr.idea.jwt.rs256.RS256Panel">
   <grid id="27dc6" binding="root" layout-manager="GridLayoutManager" row-count="2" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
